Refactor extensions class names
