+++ camel-k-core/deployment/src/main/java/org/apache/camel/k/core/quarkus/deployment/support/DeploymentSupport.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.k.core.quarkus.deployment;
+package org.apache.camel.k.core.quarkus.deployment.support;
 
 import java.util.Collection;
 import java.util.function.Function;
