+++ camel-k-cron/deployment/src/main/java/org/apache/camel/k/quarkus/cron/deployment/CronFeature.java
@@ -14,13 +14,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.k.loader.kotlin.quarkus.deployment;
+package org.apache.camel.k.quarkus.cron.deployment;
 
 import io.quarkus.deployment.annotations.BuildStep;
 import io.quarkus.deployment.builditem.FeatureBuildItem;
 
-public class Feature {
-    private static final String FEATURE = "camel-k-loader-kotlin";
+public class CronFeature {
+    private static final String FEATURE = "camel-k-cron";
 
     @BuildStep
     FeatureBuildItem feature() {
