+++ camel-k-loader-java/deployment/src/main/java/org/apache/camel/k/loader/java/quarkus/deployment/JavaSourceLoaderFeature.java
@@ -14,13 +14,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.k.core.quarkus.deployment;
+package org.apache.camel.k.loader.java.quarkus.deployment;
 
 import io.quarkus.deployment.annotations.BuildStep;
 import io.quarkus.deployment.builditem.FeatureBuildItem;
 
-public class Feature {
-    private static final String FEATURE = "camel-k-core";
+public class JavaSourceLoaderFeature {
+    private static final String FEATURE = "camel-k-loader-java";
 
     @BuildStep
     FeatureBuildItem feature() {
