+++ camel-k-loader-js/deployment/src/main/java/org/apache/camel/k/loader/js/quarkus/deployment/JavaScriptFeature.java
@@ -14,13 +14,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.k.loader.yaml.quarkus.deployment;
+package org.apache.camel.k.loader.js.quarkus.deployment;
 
 import io.quarkus.deployment.annotations.BuildStep;
 import io.quarkus.deployment.builditem.FeatureBuildItem;
 
-public class Feature {
-    private static final String FEATURE = "camel-k-loader-yaml";
+public class JavaScriptFeature {
+    private static final String FEATURE = "camel-k-loader-js";
 
     @BuildStep
     FeatureBuildItem feature() {
