+++ camel-k-loader-kotlin/deployment/src/main/java/org/apache/camel/k/loader/kotlin/quarkus/deployment/KotlinLoaderFeature.java
@@ -14,13 +14,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.k.loader.java.quarkus.deployment;
+package org.apache.camel.k.loader.kotlin.quarkus.deployment;
 
 import io.quarkus.deployment.annotations.BuildStep;
 import io.quarkus.deployment.builditem.FeatureBuildItem;
 
-public class Feature {
-    private static final String FEATURE = "camel-k-loader-java";
+public class KotlinLoaderFeature {
+    private static final String FEATURE = "camel-k-loader-kotlin";
 
     @BuildStep
     FeatureBuildItem feature() {
