+++ camel-k-loader-xml/deployment/src/main/java/org/apache/camel/k/loader/xml/quarkus/deployment/XmlLoaderFeature.java
@@ -14,13 +14,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.k.quarkus.wrap.deployment;
+package org.apache.camel.k.loader.xml.quarkus.deployment;
 
 import io.quarkus.deployment.annotations.BuildStep;
 import io.quarkus.deployment.builditem.FeatureBuildItem;
 
-public class Feature {
-    private static final String FEATURE = "camel-k-wrap";
+public class XmlLoaderFeature {
+    private static final String FEATURE = "camel-k-loader-xml";
 
     @BuildStep
     FeatureBuildItem feature() {
