+++ camel-k-loader-yaml/deployment/src/main/java/org/apache/camel/k/loader/yaml/quarkus/deployment/YamlLoaderFeature.java
@@ -14,13 +14,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.k.quarkus.deployment;
+package org.apache.camel.k.loader.yaml.quarkus.deployment;
 
 import io.quarkus.deployment.annotations.BuildStep;
 import io.quarkus.deployment.builditem.FeatureBuildItem;
 
-public class Feature {
-    private static final String FEATURE = "camel-k-runtime";
+public class YamlLoaderFeature {
+    private static final String FEATURE = "camel-k-loader-yaml";
 
     @BuildStep
     FeatureBuildItem feature() {
