+++ camel-k-runtime/deployment/src/main/java/org/apache/camel/k/quarkus/deployment/RuntimeFeature.java
@@ -14,13 +14,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.k.loader.groovy.quarkus.deployment;
+package org.apache.camel.k.quarkus.deployment;
 
 import io.quarkus.deployment.annotations.BuildStep;
 import io.quarkus.deployment.builditem.FeatureBuildItem;
 
-public class Feature {
-    private static final String FEATURE = "camel-k-loader-groovy";
+public class RuntimeFeature {
+    private static final String FEATURE = "camel-k-runtime";
 
     @BuildStep
     FeatureBuildItem feature() {
